bug reporting in navigation.
